Cleared compiler warnings
