+++ my-school/src/App.js
@@ -11,7 +11,7 @@ import Landing from './components/landingPage'
 import LoginAndRegister from './components/loginAndRegister'
 import Layout from './components/layout'
 import Dashboard from './components/dashboard'
-import Settings from './components/settings/Settings'
+import Settings from './components/settings'
 import StudentRegistration from './components/loginAndRegister/studentRegistration'
 import AddActivityForm from './components/addActivity'
 import PortfolioContainer from './components/portfolioContainer'
