+++ my-school/src/components/settings/ValidationMessage.js
@@ -1,0 +1,14 @@
+import React from "react";
+import { Text, Icon } from "@chakra-ui/core";
+
+const ValidationMessage = ({ color, message, icon }) => {
+  return (
+    <Text color={color} paddingLeft="21vw">
+      <Icon name={icon} size="24px" color={color} /> {message}
+    </Text>
+  );
+};
+
+export default ValidationMessage;
+
+//"red.500"
